takes user company input and creates an api call. Need to apppend to page and display posts
