+++ package.json
@@ -22,8 +22,10 @@
     "nodemon": "^2.0.2",
     "passport": "^0.4.1",
     "passport-local": "^1.0.0",
+    "query-string": "^6.9.0",
     "sequelize": "^5.8.6",
-    "shortid": "^2.2.15"
+    "shortid": "^2.2.15",
+    "url": "^0.11.0"
   },
   "devDependencies": {
     "chai": "^4.2.0",
