+++ public/HTML/company.html
@@ -58,7 +58,7 @@
 
 
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-    <script type="text/javascript" src="../js/members.js"></script>
+    <!-- <script type="text/javascript" src="../js/members.js"></script> -->
     <script type="text/javascript" src="../js/companyFront.js"></script>
     <script src="https://code.jquery.com/jquery-3.4.1.js"
         integrity="sha256-WpOohJOqMqqyKL9FccASB9O0KwACQJpFTUBLTYOVvVU=" crossorigin="anonymous"></script>
