+++ routes/htmlRoutes.js
@@ -28,5 +28,8 @@ module.exports = function(app) {
   app.get("/members", isAuthenticated, function(req, res) {
     res.sendFile(path.join(__dirname, "../public/HTML/members.html"));
   });
-
+  app.get("/company", isAuthenticated, function(req, res) {
+    res.sendFile(path.join(__dirname, "../public/HTML/company.html"));
+  });
+  
 };
